[MIG] hr_timesheet_activity_begin_end: Update to 15.0
