--- hr_timesheet_activity_begin_end/i18n/de.po
@@ -1,69 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* hr_timesheet_activity_begin_end
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 12.0\n"
-"Report-Msgid-Bugs-To: \n"
-"PO-Revision-Date: 2019-07-11 12:43+0000\n"
-"Last-Translator: Maria Sparenberg <maria.sparenberg@gmx.net>\n"
-"Language-Team: none\n"
-"Language: de\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=2; plural=n != 1;\n"
-"X-Generator: Weblate 3.7.1\n"
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model,name:hr_timesheet_activity_begin_end.model_account_analytic_line
-msgid "Analytic Line"
-msgstr "Kostenstellenbuchung"
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_account_analytic_line__time_start
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_hr_timesheet_switch__time_start
-msgid "Begin Hour"
-msgstr "Startzeit"
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_account_analytic_line__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_account_analytic_line__time_stop
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_hr_timesheet_switch__time_stop
-msgid "End Hour"
-msgstr "Endzeit"
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_account_analytic_line__id
-msgid "ID"
-msgstr ""
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_account_analytic_line____last_update
-msgid "Last Modified on"
-msgstr ""
-
-#. module: hr_timesheet_activity_begin_end
-#: code:addons/hr_timesheet_activity_begin_end/models/account_analytic_line.py:0
-#, python-format
-msgid "Lines can't overlap:\n"
-msgstr "Die Zeilen d├╝rfen sich nicht ├╝berschneiden:\n"
-
-#. module: hr_timesheet_activity_begin_end
-#: code:addons/hr_timesheet_activity_begin_end/models/account_analytic_line.py:0
-#, python-format
-msgid "The beginning hour (%s) must precede the ending hour (%s)."
-msgstr "Die Startzeit (%s) muss vor der Endzeit (%s) liegen."
-
-#. module: hr_timesheet_activity_begin_end
-#: code:addons/hr_timesheet_activity_begin_end/models/account_analytic_line.py:0
-#, python-format
-msgid ""
-"The duration (%s) must be equal to the difference between the hours (%s)."
-msgstr ""
-"Die Dauer (%s) muss gleich der Differenz zwischen den Stunden (%s) sein."
