--- hr_timesheet_activity_begin_end/i18n/sv.po
@@ -1,68 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* hr_timesheet_activity_begin_end
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"PO-Revision-Date: 2021-11-16 15:36+0000\n"
-"Last-Translator: Simon S <simon.stromberg@vertel.se>\n"
-"Language-Team: none\n"
-"Language: sv\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=2; plural=n != 1;\n"
-"X-Generator: Weblate 4.3.2\n"
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model,name:hr_timesheet_activity_begin_end.model_account_analytic_line
-msgid "Analytic Line"
-msgstr "Objektrad"
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_account_analytic_line__time_start
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_hr_timesheet_switch__time_start
-msgid "Begin Hour"
-msgstr "Starttid"
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_account_analytic_line__display_name
-msgid "Display Name"
-msgstr "Visningsnamn"
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_account_analytic_line__time_stop
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_hr_timesheet_switch__time_stop
-msgid "End Hour"
-msgstr "Sluttid"
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_account_analytic_line__id
-msgid "ID"
-msgstr "ID"
-
-#. module: hr_timesheet_activity_begin_end
-#: model:ir.model.fields,field_description:hr_timesheet_activity_begin_end.field_account_analytic_line____last_update
-msgid "Last Modified on"
-msgstr "Senast ├ñndrad den"
-
-#. module: hr_timesheet_activity_begin_end
-#: code:addons/hr_timesheet_activity_begin_end/models/account_analytic_line.py:0
-#, python-format
-msgid "Lines can't overlap:\n"
-msgstr "Rader f├Ñr inte ├╢verlappa varandra:\n"
-
-#. module: hr_timesheet_activity_begin_end
-#: code:addons/hr_timesheet_activity_begin_end/models/account_analytic_line.py:0
-#, python-format
-msgid "The beginning hour (%s) must precede the ending hour (%s)."
-msgstr "Starttiden (%s) m├Ñste f├╢reg├Ñ sluttimmen (%s)."
-
-#. module: hr_timesheet_activity_begin_end
-#: code:addons/hr_timesheet_activity_begin_end/models/account_analytic_line.py:0
-#, python-format
-msgid ""
-"The duration (%s) must be equal to the difference between the hours (%s)."
-msgstr "Varaktigheten (%s) m├Ñste vara lika med skillnaden mellan tiderna (%s)."
