+++ hr_timesheet_begin_end/i18n/cs_CZ.po
@@ -1,0 +1,71 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# * hr_timesheet_begin_end
+#
+# Translators:
+# Lukáš Spurný <lukasspurny8@gmail.com>, 2018
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 10.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2018-02-23 01:55+0000\n"
+"PO-Revision-Date: 2018-02-23 01:55+0000\n"
+"Last-Translator: Lukáš Spurný <lukasspurny8@gmail.com>, 2018\n"
+"Language-Team: Czech (Czech Republic) (https://www.transifex.com/oca/"
+"teams/23907/cs_CZ/)\n"
+"Language: cs_CZ\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: nplurals=3; plural=(n==1) ? 0 : (n>=2 && n<=4) ? 1 : 2;\n"
+
+#. module: hr_timesheet_begin_end
+#: model:ir.model,name:hr_timesheet_begin_end.model_account_analytic_line
+msgid "Analytic Line"
+msgstr "Analytická linka"
+
+#. module: hr_timesheet_begin_end
+#: model:ir.model.fields,field_description:hr_timesheet_begin_end.field_account_analytic_line__time_start
+#: model:ir.model.fields,field_description:hr_timesheet_begin_end.field_hr_timesheet_switch__time_start
+msgid "Begin Hour"
+msgstr "Počátek hodiny"
+
+#. module: hr_timesheet_begin_end
+#: model:ir.model.fields,field_description:hr_timesheet_begin_end.field_account_analytic_line__display_name
+msgid "Display Name"
+msgstr ""
+
+#. module: hr_timesheet_begin_end
+#: model:ir.model.fields,field_description:hr_timesheet_begin_end.field_account_analytic_line__time_stop
+#: model:ir.model.fields,field_description:hr_timesheet_begin_end.field_hr_timesheet_switch__time_stop
+msgid "End Hour"
+msgstr "Koncová hodina"
+
+#. module: hr_timesheet_begin_end
+#: model:ir.model.fields,field_description:hr_timesheet_begin_end.field_account_analytic_line__id
+msgid "ID"
+msgstr ""
+
+#. module: hr_timesheet_begin_end
+#: model:ir.model.fields,field_description:hr_timesheet_begin_end.field_account_analytic_line____last_update
+msgid "Last Modified on"
+msgstr ""
+
+#. module: hr_timesheet_begin_end
+#: code:addons/hr_timesheet_begin_end/models/account_analytic_line.py:0
+#, python-format
+msgid "Lines can't overlap:\n"
+msgstr "Linky se nemohou překrývat:\n"
+
+#. module: hr_timesheet_begin_end
+#: code:addons/hr_timesheet_begin_end/models/account_analytic_line.py:0
+#, python-format
+msgid "The beginning hour (%(html_start)s) must precede the ending hour (%(html_stop)s)."
+msgstr "Počáteční hodina (%(html_start)s) musí předcházet koncovou hodinu (%(html_stop)s)."
+
+#. module: hr_timesheet_begin_end
+#: code:addons/hr_timesheet_begin_end/models/account_analytic_line.py:0
+#, python-format
+msgid ""
+"The duration (%(html_unit_amount)s) must be equal to the difference between the hours (%(html_hours)s)."
+msgstr "Doba trvání (%(html_unit_amount)s) musí být rovna rozdílu mezi hodinami (%(html_hours)s)."
