+++ hr_timesheet_begin_end/readme/CONTRIBUTORS.rst
