+++ hr_timesheet_sheet_begin_end/__init__.py
@@ -1,0 +1,1 @@
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
