+++ hr_timesheet_sheet_begin_end/i18n/sv_SE.po
@@ -1,0 +1,54 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* hr_timesheet_sheet_begin_end
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 14.0\n"
+"Report-Msgid-Bugs-To: \n"
+"Last-Translator: Automatically generated\n"
+"Language-Team: none\n"
+"Language: sv_SE\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: nplurals=2; plural=n != 1;\n"
+
+#. module: hr_timesheet_sheet_begin_end
+#: model:ir.model,name:hr_timesheet_sheet_begin_end.model_account_analytic_line
+msgid "Analytic Line"
+msgstr ""
+
+#. module: hr_timesheet_sheet_begin_end
+#: model:ir.model.fields,field_description:hr_timesheet_sheet_begin_end.field_account_analytic_line__time_start
+#: model:ir.model.fields,field_description:hr_timesheet_sheet_begin_end.field_hr_timesheet_switch__time_start
+msgid "Begin Hour"
+msgstr ""
+
+#. module: hr_timesheet_sheet_begin_end
+#: model:ir.model.fields,field_description:hr_timesheet_sheet_begin_end.field_account_analytic_line__display_name
+msgid "Display Name"
+msgstr ""
+
+#. module: hr_timesheet_sheet_begin_end
+#: model:ir.model.fields,field_description:hr_timesheet_sheet_begin_end.field_account_analytic_line__time_stop
+#: model:ir.model.fields,field_description:hr_timesheet_sheet_begin_end.field_hr_timesheet_switch__time_stop
+msgid "End Hour"
+msgstr ""
+
+#. module: hr_timesheet_sheet_begin_end
+#: model:ir.model.fields,field_description:hr_timesheet_sheet_begin_end.field_account_analytic_line__id
+msgid "ID"
+msgstr ""
+
+#. module: hr_timesheet_sheet_begin_end
+#: model:ir.model.fields,field_description:hr_timesheet_sheet_begin_end.field_account_analytic_line____last_update
+msgid "Last Modified on"
+msgstr ""
+
+#. module: hr_timesheet_sheet_begin_end
+#: code:addons/hr_timesheet_sheet_begin_end/models/account_analytic_line.py:0
+#, python-format
+msgid "Lines can't overlap:\n"
+msgstr ""
+
