+++ hr_timesheet_sheet_begin_end/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+Adds starting and ending hours fields on the timesheet sheet views.
