+++ setup/_metapackage/setup.py
@@ -9,13 +9,16 @@ setuptools.setup(
     version=version,
     install_requires=[
         'odoo-addon-hr_employee_product>=15.0dev,<15.1dev',
+        'odoo-addon-hr_timesheet_begin_end>=15.0dev,<15.1dev',
         'odoo-addon-hr_timesheet_sheet>=15.0dev,<15.1dev',
+        'odoo-addon-hr_timesheet_sheet_begin_end>=15.0dev,<15.1dev',
         'odoo-addon-hr_timesheet_task_domain>=15.0dev,<15.1dev',
         'odoo-addon-hr_timesheet_task_required>=15.0dev,<15.1dev',
         'odoo-addon-hr_timesheet_task_stage>=15.0dev,<15.1dev',
         'odoo-addon-hr_timesheet_time_type>=15.0dev,<15.1dev',
         'odoo-addon-sale_timesheet_line_exclude>=15.0dev,<15.1dev',
         'odoo-addon-sale_timesheet_task_exclude>=15.0dev,<15.1dev',
+
     ],
     classifiers=[
         'Programming Language :: Python',
