--- setup/hr_timesheet_activity_begin_end/odoo/addons/hr_timesheet_activity_begin_end
@@ -1,1 +1,0 @@
-../../../../hr_timesheet_activity_begin_end
