+++ setup/hr_timesheet_sheet_begin_end/odoo/addons/hr_timesheet_sheet_begin_end
@@ -1,0 +1,1 @@
+../../../../hr_timesheet_sheet_begin_end
